弾の表示時間をintからfloatに変更

+++ 3D_Base/Source/GameObject/StaticMeshObject/Shot/CShot.h
@@ -35,7 +35,7 @@ protected:
 		bool		m_ShotFlag;				// 発射フラグ
 		D3DXVECTOR3 m_MoveDirection;		// 移動方向
 		float		m_MoveSpeed;			// 移動速度
-		int			m_DisplayTime;			// 約3秒くらい表示する
+		float		m_DisplayTime;			// 約3秒くらい表示する
 		float		m_Gravity;				// 弾の重力
 		float		m_Velocity;				// 加速度
 	};
